Split utils tests into three files Moved all tests from other to new files
